chat_agent/pyproject.toml
[project]
name = "chat_agent"
version = "0.1.0"
description = "chat_agent using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.102.0,<1.0.0",
    "litellm>=1.60.2",
    "python-dotenv>=1.0.1",
]

[project.scripts]
chat_agent = "chat_agent.main:run"
run_crew = "chat_agent.main:kickoff"
plot = "chat_agent.main:plot"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
